integration_tests/test-requirements-for-tox.txt
https://github.com/wazo-platform/mockserver-client-python/archive/5.6.X.zip
https://github.com/wazo-platform/xivo-bus/archive/master.zip
https://github.com/wazo-platform/wazo-test-helpers/archive/master.zip
https://github.com/wazo-platform/xivo-lib-python/archive/master.zip

docker-compose
openapi-spec-validator
pyhamcrest
# pytest 4.2.0 has a bug: https://github.com/pytest-dev/pytest/issues/4700
pytest!=4.2.0
python-ldap
sh
mock

# for database tests
psycopg2-binary
sqlalchemy==1.2.18
unidecode==1.0.23
cheroot==6.5.4
kombu
stevedore
flask-cors==3.0.7
flask-restful==0.3.7
flask==1.0.2
marshmallow==3.0.0b14
werkzeug==0.14.1
https://github.com/wazo-platform/wazo-lib-rest-client/archive/master.zip
https://github.com/wazo-platform/wazo-auth-client/archive/master.zip
https://github.com/wazo-platform/wazo-confd-client/archive/master.zip
https://github.com/wazo-platform/wazo-dird-client/archive/master.zip
